Support G6110 1. Compatible with 64-bit platform 2. Fix hwc_getDisplayAttributes bug for framebuffer_test. 3. Fix format error of Live Wallpaper 4. Use sort_area_by_pos instead of sort_area_by_ypos & sort_area_by_xpos 5. Support FBDC of GPU composer case. 6. Temporary solution:        1) Close rga  (G6110_RM_RGA  1)        2) Close USE_QUEUE_DDRFREQ  (USE_QUEUE_DDRFREQ 0)
Depend on:
@hardware/libhardware/include/hardware/rga.h
commit: Update rga.h for compatible with 64bit platform

+++ hwc_rga.cpp
@@ -405,7 +405,7 @@ int set_rga_cfg(hwc_cfg_t  *cfg)
   //src cfg
   if (cfg->rga_fbAddr==0)
   {
-	   g_rga_cfg.rga_cfg.src.yrgb_addr = (unsigned int )(SRC_HANDLE_BASE);
+	   g_rga_cfg.rga_cfg.src.yrgb_addr = (unsigned long )(SRC_HANDLE_BASE);
   }
   else
   {
@@ -423,7 +423,7 @@ int set_rga_cfg(hwc_cfg_t  *cfg)
   //dst cfg
   //memset((void*)cfg->dst_handle->base,0,cfg->dst_handle->width*cfg->dst_handle->height*4);
 
-  g_rga_cfg.rga_cfg.dst.yrgb_addr = (unsigned int)(DST_HANDLE_BASE);
+  g_rga_cfg.rga_cfg.dst.yrgb_addr = (unsigned long)(DST_HANDLE_BASE);
 
   g_rga_cfg.rga_cfg.dst.vir_w = WIDTH(cfg->dst_rect);
   g_rga_cfg.rga_cfg.dst.vir_h = HEIGHT(cfg->dst_rect);
